backend/requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.2.1,<3.0.0
python-multipart>=0.0.9,<1.0.0
sqlalchemy>=2.0.30,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0
neo4j>=5.20.0,<6.0.0
redis>=5.0.4,<6.0.0
scikit-learn>=1.4.2,<2.0.0
numpy>=1.26.4,<3.0.0
python-dateutil>=2.9.0,<3.0.0
PyYAML>=6.0.1,<7.0.0
httpx>=0.27.0,<1.0.0
pytest>=8.2.0,<9.0.0
pytest-asyncio>=0.23.6,<1.0.0
aiofiles>=23.2.1,<25.0.0